Change the way how abandoned indexes are handled
Before we enforced a single active import and denied any new start. This
could lead to the situation, that an abandoned index (never done or
canceled due to some 'interruptions) blocked a new indexation for ever.

Now we allow a new start and simply log a warning for abandoned indexes.
The final call of 'done' will switch to the according index.
Additionately we check (async) if the abandoned indexes are older than 1h
and delete them if so.