Add more resilient way to fetch hero products
Previously we expected the products to appear in "chunks" according to the boost values.
However if there are rescorers in place, the whole order might be reshuffled.
The best what we can do here, is to pick the hero products from the returned result
and put them back into the hero-slices.

This improves the solution, but it doesn't solve it completely.
Maybe separate requests are necessary to guarantee the requested order.