Implement idPath filtering for path facets
Until now a path-facet parameter would eiter be
    category=Foo/Bar
or
    category.id=3

With the facet config property "ShowUnselectedOptions" set to false
this information was also used to filter select the according
path in the facet as well (and remove the other options).

But it was not possible to show a specific path in the facets using IDs.
That is only be relevant, if the same category(Id) could be placed at
different paths, e.g "Foo=1/Bar=3" and "Blub=2/Bar=3".

Now it's possible to use the id-filter with a full id-path:
    category.id=/1/3

The trailing slash is essential, because only then the value is treated as ID-path.
In case there are actually slashes inside the IDs - good luck, that's not handled yet.

I mean, seriously, who would do such stuff. :-)