Add explicit langCode parameter for putDocuments
The putDocuments has the ability to create an index if it does not exist yet.
However before the index name required a language code as a suffix, which in turn was not removed from the final index-name, which might be desired.
To avoid that, the language code is now an optional parameter for the put request.