QuerqyQueryExpander: restore normalized terms
the transforming parser normalizes the original query terms, to that the original terms are lost.
this is a problem when we actually want Elasticsearch to do the normalization (properly),
e.g. asciify should be done after stemming, otherwise we have different stemming results.
Therefore we restore the original terms here again.
However that is not that easy, since rules can delete terms or add terms.
That's why we link every original term with the normalized term with a runnable which is executed later.